Remove Rails 5.0/5.1 support and add Rails 6.1
Why is this change needed?
--------------------------
I'm working on bringing this repo up to speed to make a new release. Rails 5.1 hasn't been supported in almost 2.5 years, so it seems straightforward to stop supporting it here.

How does it address the issue?
------------------------------
- Remove Rails 5.0 and 5.1 from the Appraisals file.
- Add Rails 6.1 to the Appraisals file

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------