Add support for Rails 7, drop 5.2 and 6.0
Why is this change needed?
--------------------------
Basic gem maintenance. Those versions of Rails are no longer maintained.

How does it address the issue?
------------------------------
There's a lot of stuff here, so I'll just go file by file

- .rubocop.yml
Rails 7 requires Ruby >= 2.7 which necessitated an upgrade of Rubocop,
hence the change in RubyTargetVersion. I disabled SuggestExtensions
because it kept printing a message about the rake extension, it was
bothering me, and I didn't have time to decide if that's a thing I
wanted.

- Appraisals
Drop 5.2 and 6.0. Add Rails 7

- Gemfile.lock
Bump AAA to version 2.1 and update the rest of the gems to make sure
the library is staying up to date. In particular, dependabot said
rexml needed an update to avoid a security issue and rubocop needed an
update for reasons explained above.

- acts_as_archival.gemspec
The mfa requirement was suggested by Rubocop and it seems like a good
idea.

- gemfiles/rails_6.0.gemfile -> gemfiles/rails_7.0.gemfile
This wasn't really renamed, it's just Git mistaking a
delete-and-recreate for a rename.

- lib/acts_as_archival/version.rb
Release bump

- lib/expected_behavior/acts_as_archival.rb
All of these changes were at the behest of Rubocop. They seemed
reasonable to me.

- test/callbacks_test.rb, test/fixtures/callback_archival_4.rb -> test/fixtures/callback_archival.rb, test/fixtures/callback_archival_5.rb, test/schema.rb
I found some more things relating to Rails 4 and 5 that I missed
removing. In this case, the tests acted differently depending on
whether or not ApplicationRecord existed. All of the supported
versions of Rails now have ApplicationRecord, so it's not necessary to
have a conditional or two have different fixtures.

- test/scope_test.rb
Removed another unnecessary conditional for old Rails versions

- test/test_helper.rb
The interpolation changes were requested by Rubocop. I removed the
RAILS_4 and RAILS_5 fixture differentiators because all of the
supported Rails versions can use the same fixtures right now.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------

Co-authored-by: Tony Dewan <tony@expectedbehavior.com>